bsl-op-forbidden-overload: Fix template bug
Instead of matching on CXXOperatorCallExpr's with an overload
name of "&", match on Decls with the same name and only one
argument. This ensures that any overloaded operator& does not
have 1 argument, which effectively prevents address-of from
being overloaded.

This fixes a false positive in template code that generates
a dependent CXXOperatorCallExpr node that doesn't get resolved
until later, with (seemingly) no way to distinguish address-of
and bitwise &.